Add new tests for WysiwygComposer
